pyproject.toml
[project]
name = "cellmech"
version = "1.1.0"
description = ""
authors = [
    {name = "Utkarsh"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "seaborn (>=0.13.2,<0.14.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pyqt6 (>=6.10.0,<7.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1.0"

